+++ datumaro/cli/commands/explore.py
@@ -4,14 +4,13 @@
 
 import argparse
 import logging as log
+import os
 import os.path as osp
-
-import numpy as np
+import shutil
 
 from datumaro.components.errors import ProjectNotFoundError
 from datumaro.components.explorer import Explorer
-from datumaro.components.visualizer import Visualizer
-from datumaro.util.image import save_image
+from datumaro.util import str_to_bool
 from datumaro.util.scope import scope_add, scoped
 
 from ..util import MultilineFormatter
@@ -42,10 +41,7 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         formatter_class=MultilineFormatter,
     )
 
-    parser.add_argument(
-        "_positionals", nargs=argparse.REMAINDER, help=argparse.SUPPRESS
-    )  # workaround for -- eaten by positionals
-    parser.add_argument("target", nargs="+", default="project", help="Target dataset")
+    parser.add_argument("target", nargs="?", help="Target dataset")
     parser.add_argument(
         "-q",
         "--query",
@@ -61,9 +57,25 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         help="Directory of the project to operate on (default: current dir)",
     )
     parser.add_argument(
-        "-s", "--save", dest="save", default=True, help="Save explorer result as png"
+        "-s",
+        "--save",
+        action="store_true",
+        default=False,
+        help="Save explorer result files on explore_result folder",
     )
-
+    parser.add_argument(
+        "--stage",
+        type=str_to_bool,
+        default=True,
+        help="""
+            Include this action as a project build step.
+            If true, this operation will be saved in the project
+            build tree, allowing to reproduce the resulting dataset later.
+            Applicable only to main project targets (i.e. data sources
+            and the 'project' target, but not intermediate stages)
+            (default: %(default)s)
+            """,
+    )
     parser.set_defaults(command=explore_command)
 
     return parser
@@ -75,7 +87,7 @@ def get_sensitive_args():
             "target",
             "query",
             "topk",
-            "save",
+            "project_dir",
         ]
     }
 
@@ -89,36 +101,61 @@ def explore_command(args):
         if args.project_dir:
             raise
 
-    dataset, _ = parse_full_revpath(args.target[0], project)
+    if args.target:
+        targets = [args.target]
+    else:
+        targets = list(project.working_tree.sources)
 
-    explorer = Explorer(dataset)
+    source_datasets = []
+    for target in targets:
+        target_dataset, _ = parse_full_revpath(target, project)
+        source_datasets.append(target_dataset)
+
+    explorer_args = {"save_hashkey": True}
+    build_tree = project.working_tree.clone()
+    for target in targets:
+        build_tree.build_targets.add_explore_stage(target, params=explorer_args)
+
+    explorer = Explorer(*source_datasets)
+    for dataset in source_datasets:
+        dst_dir = dataset.data_path
+        dataset.save(dst_dir, save_media=True)
+
+    if args.stage:
+        project.working_tree.config.update(build_tree.config)
+        project.working_tree.save()
 
     # Get query datasetitem through query path
     if osp.exists(args.query):
-        query_datasetitem = dataset.get_datasetitem_by_path(args.query)
+        query_datasetitem = None
+        for dataset in source_datasets:
+            try:
+                query_datasetitem = dataset.get_datasetitem_by_path(args.query)
+            except Exception:
+                continue
+            if not query_datasetitem:
+                break
     else:
         query_datasetitem = args.query
 
     results = explorer.explore_topk(query_datasetitem, args.topk)
 
-    subset_list = []
-    id_list = []
     result_path_list = []
-    log.info("Most similar {} results of query in dataset".format(args.topk))
+    log.info(f"Most similar {args.topk} results of query in dataset")
     for result in results:
-        subset_list.append(result.subset)
-        id_list.append(result.id)
         path = getattr(result.media, "path", None)
         result_path_list.append(path)
-        log.info("id: {} | subset: {} | path : {}".format(result.id, result.subset, path))
-
-    visualizer = Visualizer(dataset, figsize=(20, 20), alpha=0)
-    fig = visualizer.vis_gallery(id_list, subset_list)
+        log.info(f"id: {result.id} | subset: {result.subset} | path : {path}")
 
     if args.save:
-        fig.canvas.draw()
-        data = np.frombuffer(fig.canvas.tostring_rgb(), dtype=np.uint8)
-        data = data.reshape(fig.canvas.get_width_height()[::-1] + (3,))
-        save_image(osp.join("./explorer.png"), data, create_dir=True)
+        saved_result_path = osp.join(args.project_dir, "explore_result")
+        if osp.exists(saved_result_path):
+            shutil.rmtree(saved_result_path)
+        os.makedirs(saved_result_path)
+        for result in results:
+            saved_subset_path = osp.join(saved_result_path, result.subset)
+            if not osp.exists(saved_subset_path):
+                os.makedirs(saved_subset_path)
+            shutil.copyfile(path, osp.join(saved_subset_path, result.id + ".jpg"))
 
     return 0