+++ datumaro/components/project.py
@@ -288,6 +288,7 @@ class BuildStageType(Enum):
     filter = auto()
     convert = auto()
     inference = auto()
+    explore = auto()
 
 
 class Pipeline:
@@ -910,6 +911,18 @@ class ProjectBuildTargets(CrudProxy[BuildTarget]):
             {
                 "type": BuildStageType.convert.name,
                 "kind": format,
+                "params": params or {},
+            },
+            name=name,
+        )
+
+    def add_explore_stage(
+        self, target: str, params: Optional[Dict] = None, name: Optional[str] = None
+    ):
+        return self.add_stage(
+            target,
+            {
+                "type": BuildStageType.explore.name,
                 "params": params or {},
             },
             name=name,