+++ datumaro/plugins/data_formats/datumaro/base.py
@@ -5,12 +5,15 @@
 import os.path as osp
 from typing import Optional
 
+import numpy as np
+
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
     Caption,
     Cuboid3d,
     Ellipse,
+    HashKey,
     Label,
     LabelCategories,
     MaskCategories,
@@ -25,6 +28,7 @@ from datumaro.components.errors import DatasetImportError, MediaTypeError
 from datumaro.components.importer import ImportContext
 from datumaro.components.media import Image, MediaElement, PointCloud
 from datumaro.util import parse_json_file
+from datumaro.util.meta_file_util import has_hashkey_file, parse_hashkey_file
 from datumaro.version import __version__
 
 from .format import DATUMARO_FORMAT_VERSION, DatumaroPath
@@ -99,7 +103,8 @@ class DatumaroBase(SubsetBase):
         """Actual implementation of loading Datumaro format."""
         self._infos = self._load_infos(self._parsed_anns)
         self._categories = self._load_categories(self._parsed_anns)
-        self._items = self._load_items(self._parsed_anns)
+        items = self._load_items(self._parsed_anns)
+        self._items = self._load_hash_key(items)
 
     @staticmethod
     def _load_infos(parsed):
@@ -147,6 +152,7 @@ class DatumaroBase(SubsetBase):
 
     def _load_items(self, parsed):
         items = []
+
         for item_desc in parsed["items"]:
             item_id = item_desc["id"]
 
@@ -202,6 +208,16 @@ class DatumaroBase(SubsetBase):
 
             items.append(item)
 
+        return items
+
+    def _load_hash_key(self, items):
+        if not has_hashkey_file(self._rootpath):
+            return items
+
+        hashkey_dict = parse_hashkey_file(self._rootpath)
+        for item in items:
+            hash_key = hashkey_dict[item.subset + "/" + item.id]
+            item.annotations.append(HashKey(hash_key=np.asarray(hash_key, dtype=np.uint8)))
         return items
 
     @staticmethod
@@ -317,6 +333,8 @@ class DatumaroBase(SubsetBase):
                     )
                 )
 
+            elif ann_type == AnnotationType.hash_key:
+                continue
             else:
                 raise NotImplementedError()