+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/973>)
 - Enrich stack trace for better user experience when importing
   (<https://github.com/openvinotoolkit/datumaro/pull/992>)
+- Save and load hashkey for explorer
+  (<https://github.com/openvinotoolkit/datumaro/pull/981>)
 
 ### Bug fixes
 - Fix Mapillary Vistas data format